refactor(k8s): move internal/chaosengine/client → platform/k8s/chaosclient
Directory rename requires package declaration update: client → chaosclient.
