+++ aegislab/src/internal/chaosengine/resourcelookup/lookup.go
@@ -15,7 +15,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 
-	"aegis/internal/chaosengine/client"
+	"aegis/platform/k8s/chaosclient"
 	"aegis/platform/systemconfig"
 )
 
@@ -209,7 +209,7 @@ func (s *systemCache) GetAllAppLabels(ctx context.Context, namespace string, key
 		return labels, nil
 	}
 
-	labels, err := client.GetLabels(ctx, namespace, key)
+	labels, err := chaosclient.GetLabels(ctx, namespace, key)
 	if err != nil {
 		return nil, err
 	}
@@ -366,7 +366,7 @@ func (s *systemCache) GetAllContainers(ctx context.Context, namespace string) ([
 		}
 	}
 
-	containers, err := client.GetContainersWithAppLabel(ctx, namespace, systemconfig.GetAppLabelKey(s.system))
+	containers, err := chaosclient.GetContainersWithAppLabel(ctx, namespace, systemconfig.GetAppLabelKey(s.system))
 	if err != nil {
 		return nil, err
 	}
