+++ aegislab/src/platform/k8s/chaosclient/kubernetes.go
@@ -1,4 +1,4 @@
-package client
+package chaosclient
 
 import (
 	"context"
