+++ aegislab/src/platform/k8s/chaosclient/kubernetes_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package client
+package chaosclient
 
 import (
 	"context"
