+++ aegislab/src/platform/k8s/chaosclient/kubernetes_test.go
@@ -1,4 +1,4 @@
-package client
+package chaosclient
 
 import (
 	"context"
